'wip issue homework

+++ frontend/src/app/modules/homework/pages/issue-homework/issue-homework.service.adapter.ts
@@ -1,4 +1,5 @@
 
+import { Homework } from 'app/services/modules/homework/models/homework';
 import { IssueHomeworkComponent } from './issue-homework.component';
 
 export class IssueHomeworkServiceAdapter {
@@ -53,6 +54,7 @@ export class IssueHomeworkServiceAdapter {
             this.vm.homeworkService.getObjectList(this.vm.homeworkService.homeworks, {parentClassSubject: this.vm.selectedSubject.classSubjectDbId}),
             this.vm.homeworkService.getObjectList(this.vm.homeworkService.homework_question, homework_data),
         ]).then(value =>{
+            this.vm.homeworkImagesList = value[1];
             this.initialiseHomeworks(value[0], value[1]);        
             this.vm.isLoading = false;
         },error =>{
@@ -69,6 +71,7 @@ export class IssueHomeworkServiceAdapter {
                 let tempData = {
                     id: currentHomework.id,
                     homeworkName: currentHomework.homeworkName,
+                    parentClassSubject: currentHomework.parentClassSubject,
                     startDate: currentHomework.startDate,
                     startTime: currentHomework.startTime,
                     endDate: currentHomework.endDate,
@@ -91,6 +94,10 @@ export class IssueHomeworkServiceAdapter {
 
 
     deleteHomework(homeworkId: any): any{
+
+        if(!confirm("Are you sure you want to delete this homework?")) {
+            return;
+        }
         this.vm.isLoading = true;
         this.vm.homeworkService.deleteObject(this.vm.homeworkService.homeworks, {id: homeworkId}).then(value =>{
             alert('Homework Deleted')
@@ -101,5 +108,63 @@ export class IssueHomeworkServiceAdapter {
         };
     }
 
-    
+    updateHomework(data :any): any{
+
+        let previousHomework = this.vm.homeworkList.find(homework => homework.id === data.id);
+
+        let tempHomeworkData = {
+            id: data.id,
+            homeworkName: data.homeworkName ,
+            parentClassSubject: data.parentClassSubject,
+            startDate: data.startDate,
+            startTime: data.startTime,
+            endDate: data.endDate,
+            endTime: data.endTime,
+            homeworkText: data.homeworkText,
+        }
+
+        const promises = [];
+
+        let index = 0;
+        promises.push(this.vm.homeworkService.updateObject(this.vm.homeworkService.homeworks, tempHomeworkData));
+        data.homeworkImages.forEach(image =>{
+            let temp = previousHomework.homeworkImages.find(images => images === image);
+            if(temp === undefined){
+                let tempData = {
+                    parentHomework: data.id,
+                    questionImage: image,
+                }
+                let temp_form_data = new FormData();
+                const layout_data = { ...tempData,};
+                Object.keys(layout_data).forEach(key => {
+                    if (key === 'questionImage' ) {
+                        temp_form_data.append(key, this.vm.dataURLtoFile(layout_data[key], 'questionImage' + index +'.jpeg'));
+                    } else {
+                        temp_form_data.append(key, layout_data[key]);
+                    }
+                });
+                index = index + 1;
+                promises.push(this.vm.homeworkService.createObject(this.vm.homeworkService.homework_question, temp_form_data));
+            }
+        });
+
+        previousHomework.homeworkImages.forEach(image =>{
+            let temp = data.homeworkImages.find(images => images === image);
+            if(temp === undefined){
+                let tempId = this.vm.homeworkImagesList.find(images => images.questionImage == image).id;
+                console.log(tempId);
+                promises.push(this.vm.homeworkService.deleteObject(this.vm.homeworkService.homework_question,{'id': tempId}));
+            }
+        });
+
+        console.log(tempHomeworkData);
+        Promise.all(promises).then(value =>{
+            this.getHomeworks();
+            alert('Homework Edited Successfully');
+            this.vm.isLoading = false;
+        }),error =>{
+            this.vm.isLoading = false;
+        }
+        
+    }
 }
